% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_blueprint.R
\name{blueprint_discrete_palette}
\alias{blueprint_discrete_palette}
\title{Return Blueprint discrete palette}
\usage{
blueprint_discrete_palette()
}
\value{

}
\description{
Return Blueprint discrete palette
}
