% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_blueprint.R
\name{blueprint_discrete_palette_bold}
\alias{blueprint_discrete_palette_bold}
\title{Return Blueprint discrete palette bold}
\usage{
blueprint_discrete_palette_bold()
}
\value{

}
\description{
Return Blueprint discrete palette bold
}
