% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint_pal.R
\name{blueprint_pal}
\alias{blueprint_pal}
\title{blueprint_pal()}
\usage{
blueprint_pal(
  palette = primary_brand_colours,
  primary = palette[1],
  other = palette[2],
  direction = 1,
  ...
)
}
\arguments{
\item{palette}{The approved Blueprint palette of choice. See 'blueprint_palettes'.}

\item{primary}{Lets you manually specify the colour of the first level of a two-level palette. Default is the first colour in the chosen palette.}

\item{other}{Lets you manually specify the colour of the second level of two-level palette. Default is the second colour in the chosen palette.}

\item{direction}{Sets the order of colors in the scale. If 1, the default, colors are ordered from darkest to lightest. If -1, the order of colors is reversed.}

\item{...}{Other arguments passed on to discrete_scale(), continuous_scale(), or binned_scale to control name, limits, breaks, labels and so forth.}

\item{scale_name}{A string to serve as the label of the colour scale as displayed on the plot.}

\item{aesthetics}{Character string or vector of character strings listing the name(s) of the aesthetic(s) that this scale works with. This can be useful, for example, to apply colour settings to the colour and fill aesthetics at the same time, via aesthetics = c("colour", "fill").}

\item{space}{Required by continuous_scale.}

\item{na.value}{Required by continuous_scale.}

\item{guide}{Required by continuous_scale.}

\item{values}{Required by continuous_scale.}
}
\description{
The blueprint_pal() function, to return a vector of colours to be fed to discrete_scale() or continuous_scale()
}
