% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_blueprint_palettes.R
\name{display_blueprint_palettes}
\alias{display_blueprint_palettes}
\title{display_blueprint_all}
\usage{
display_blueprint_palettes()
}
\description{
This function displays all of the Blueprint palettes
drawn from the official corporate style guide
}
