% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_blueprint.R
\name{load_fonts}
\alias{load_fonts}
\title{Return Blueprint discrete palette}
\usage{
load_fonts()
}
\value{



}
\description{
Return Blueprint discrete palette
}
\examples{
Return Blueprint discrete palette bold

}
