% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_blueprint.R
\name{Blueprint Colour Scales}
\alias{scale_colour_blueprint_d}
\alias{scale_fill_blueprint_d}
\alias{scale_colour_blueprint_c}
\alias{scale_fill_blueprint_c}
\title{Blueprint Colour Scales}
\usage{
scale_colour_blueprint_d(
  ...,
  scale_name = "blueprint_d",
  direction = 1,
  palette = primary_brand_colours,
  aesthetics = "colour",
  primary = palette[1],
  other = palette[2]
)

scale_fill_blueprint_d(
  ...,
  scale_name = "blueprint_d",
  direction = 1,
  palette = primary_brand_colours,
  aesthetics = "fill",
  primary = palette[1],
  other = palette[2]
)

scale_colour_blueprint_c(
  ...,
  direction = 1,
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = "colourbar",
  aesthetics = "colour",
  palette = primary_brand_colours
)

scale_fill_blueprint_c(
  ...,
  direction = 1,
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = "colourbar",
  aesthetics = "fill",
  palette = primary_brand_colours,
  primary = palette[1],
  other = palette[2]
)
}
\arguments{
\item{...}{Other arguments passed on to discrete_scale(), continuous_scale(), or binned_scale to control name, limits, breaks, labels and so forth.}

\item{scale_name}{A string to serve as the label of the colour scale as displayed on the plot.}

\item{direction}{Sets the order of colors in the scale. If 1, the default, colors are ordered from darkest to lightest. If -1, the order of colors is reversed.}

\item{palette}{The approved Blueprint palette of choice. See 'blueprint_palettes'.}

\item{aesthetics}{Character string or vector of character strings listing the name(s) of the aesthetic(s) that this scale works with. This can be useful, for example, to apply colour settings to the colour and fill aesthetics at the same time, via aesthetics = c("colour", "fill").}

\item{primary}{Lets you manually specify the colour of the first level of a two-level palette. Default is the first colour in the chosen palette.}

\item{other}{Lets you manually specify the colour of the second level of two-level palette. Default is the second colour in the chosen palette.}

\item{values}{Required by continuous_scale.}

\item{space}{Required by continuous_scale.}

\item{na.value}{Required by continuous_scale.}

\item{guide}{Required by continuous_scale.}
}
\description{
Blueprint colour scales based on the corporate style guide
}
\concept{colour scales}
