% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_blueprint.R
\name{scale_fill_blueprint}
\alias{scale_fill_blueprint}
\title{Fill continuous or discrete mappings with Blueprint palettes}
\usage{
scale_fill_blueprint(discrete = FALSE, palette = "default", ...)
}
\arguments{
\item{palette}{discrete palette to be used}

\item{...}{}
}
\value{

}
\description{
Fill continuous or discrete mappings with Blueprint palettes
}
